Suggestions: - Added test for Greeting - Move XDate Exception to domain package (and the corresponding test) - Moved Fakes to test->domain->ports
