+++ src/main/kotlin/it/xpug/kata/birthday_greetings/domain/InvalidInputDateException.kt
@@ -1,3 +1,3 @@
-package it.xpug.kata.birthday_greetings.adapters.inbound
+package it.xpug.kata.birthday_greetings.domain
 
 class InvalidInputDateException(dateInput: String) : RuntimeException("Invalid date format: '$dateInput'. Please provide a date in the format YYYY/MM/DD.")
