+++ src/test/kotlin/it/xpug/kata/birthday_greetings/domain/XDateTest.kt
@@ -1,6 +1,8 @@
 package it.xpug.kata.birthday_greetings.domain
 
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
 
 class XDateTest {
@@ -28,5 +30,17 @@ class XDateTest {
         assertThat(date.isSameDay(notSameMonth))
             .`as`("not same month")
             .isFalse()
+    }
+
+    @Test
+    fun `throws exception when invalid date format is provided`() {
+        // given
+        val invalidDateFormat = "30-03-2025" // Should be YYYY/MM/DD
+
+        // when & then
+        val exception = assertThrows(InvalidInputDateException::class.java) {
+            XDate(invalidDateFormat)
+        }
+        assertEquals("Invalid date format: '30-03-2025'. Please provide a date in the format YYYY/MM/DD.", exception.message)
     }
 }
