+++ src/test/kotlin/it/xpug/kata/birthday_greetings/domain/ports/outbound/InMemoryEmployeeRepository.kt
@@ -1,7 +1,6 @@
-package it.xpug.kata.birthday_greetings.adapters.outbound
+package it.xpug.kata.birthday_greetings.domain.ports.outbound
 
 import it.xpug.kata.birthday_greetings.domain.Employee
-import it.xpug.kata.birthday_greetings.domain.ports.outbound.EmployeeRepository
 
 class InMemoryEmployeeRepository : EmployeeRepository {
